# Improve performance of segment querying endpoints