This commit addresses the excessive amount of database queries being done when querying segments. Previously, one query was done per category. With this PR one query is done for all categories and the returned data is processed on the server itself.

Additionally, when querying by hash, the segments are queried by hash instead of querying for video ids by hash and then querying segments by video id. This means that the API behavior changes slightly and one of the tests was adjusted for it. The old behavior was, in my opinion, pretty useless anyway, so I don't think this is a problem.

Benchmarks done using autocannon on a version of the production database from earlier today are below. The results fluctuated a bit between runs, but the improvement is still plainly visible.

## `/api/skipSegments`

### Old code

```
Running 10s test @ http://localhost:8080/api/skipSegments?categories=%5B%22sponsor%22,%22intro%22,%22outro%22,%22interaction%22,%22selfpromo%22,%22music_offtopic%22%5D&videoID=165bpjQ2ZNs
1 connections

┌─────────┬──────┬──────┬───────┬──────┬────────┬────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg    │ Stdev  │ Max   │
├─────────┼──────┼──────┼───────┼──────┼────────┼────────┼───────┤
│ Latency │ 0 ms │ 0 ms │ 1 ms  │ 2 ms │ 0.2 ms │ 1.3 ms │ 73 ms │
└─────────┴──────┴──────┴───────┴──────┴────────┴────────┴───────┘
┌───────────┬────────┬────────┬─────────┬─────────┬─────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%   │ Avg     │ Stdev  │ Min    │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 871    │ 871    │ 1245    │ 1518    │ 1242.3  │ 193.33 │ 871    │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 888 kB │ 888 kB │ 1.27 MB │ 1.55 MB │ 1.27 MB │ 197 kB │ 888 kB │
└───────────┴────────┴────────┴─────────┴─────────┴─────────┴────────┴────────┘

Req/Bytes counts sampled once per second.

12k requests in 10.02s, 12.7 MB read
```

### New code

```
Running 10s test @ http://localhost:8080/api/skipSegments?categories=%5B%22sponsor%22,%22intro%22,%22outro%22,%22interaction%22,%22selfpromo%22,%22music_offtopic%22%5D&videoID=165bpjQ2ZNs
1 connections

┌─────────┬──────┬──────┬───────┬──────┬─────────┬─────────┬───────┐
│ Stat    │ 2.5% │ 50%  │ 97.5% │ 99%  │ Avg     │ Stdev   │ Max   │
├─────────┼──────┼──────┼───────┼──────┼─────────┼─────────┼───────┤
│ Latency │ 0 ms │ 0 ms │ 1 ms  │ 1 ms │ 0.07 ms │ 0.66 ms │ 28 ms │
└─────────┴──────┴──────┴───────┴──────┴─────────┴─────────┴───────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev  │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼────────┼─────────┤
│ Req/Sec   │ 1538    │ 1538    │ 1972    │ 2079    │ 1945.46 │ 138.77 │ 1538    │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼────────┼─────────┤
│ Bytes/Sec │ 1.57 MB │ 1.57 MB │ 2.01 MB │ 2.12 MB │ 1.98 MB │ 141 kB │ 1.57 MB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴────────┴─────────┘

Req/Bytes counts sampled once per second.

21k requests in 11.01s, 21.8 MB read
```

## `/api/skipSegments/:prefix`

### Old code

```
Running 10s test @ http://localhost:8080/api/skipSegments/aaaa?categories=%5B%22sponsor%22,%22intro%22,%22outro%22,%22interaction%22,%22selfpromo%22,%22music_offtopic%22%5D
1 connections

┌─────────┬────────┬────────┬────────┬────────┬───────────┬──────────┬────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%  │ 99%    │ Avg       │ Stdev    │ Max    │
├─────────┼────────┼────────┼────────┼────────┼───────────┼──────────┼────────┤
│ Latency │ 507 ms │ 548 ms │ 682 ms │ 682 ms │ 565.48 ms │ 44.14 ms │ 682 ms │
└─────────┴────────┴────────┴────────┴────────┴───────────┴──────────┴────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg    │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼────────┼─────────┼─────────┤
│ Req/Sec   │ 1       │ 1       │ 2       │ 2       │ 1.73   │ 0.45    │ 1       │
├───────────┼─────────┼─────────┼─────────┼─────────┼────────┼─────────┼─────────┤
│ Bytes/Sec │ 3.07 kB │ 3.07 kB │ 6.14 kB │ 6.14 kB │ 5.3 kB │ 1.37 kB │ 3.07 kB │
└───────────┴─────────┴─────────┴─────────┴─────────┴────────┴─────────┴─────────┘

Req/Bytes counts sampled once per second.

19 requests in 11.01s, 58.3 kB read
```

### New code

```
Running 10s test @ http://localhost:8080/api/skipSegments/aaaa?categories=%5B%22sponsor%22,%22intro%22,%22outro%22,%22interaction%22,%22selfpromo%22,%22music_offtopic%22%5D
1 connections

┌─────────┬───────┬───────┬────────┬────────┬──────────┬─────────┬────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5%  │ 99%    │ Avg      │ Stdev   │ Max    │
├─────────┼───────┼───────┼────────┼────────┼──────────┼─────────┼────────┤
│ Latency │ 88 ms │ 95 ms │ 102 ms │ 103 ms │ 94.43 ms │ 3.37 ms │ 105 ms │
└─────────┴───────┴───────┴────────┴────────┴──────────┴─────────┴────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 10      │ 10      │ 10      │ 11      │ 10.5    │ 0.5     │ 10      │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 30.7 kB │ 30.7 kB │ 30.7 kB │ 33.8 kB │ 32.2 kB │ 1.54 kB │ 30.7 kB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘

Req/Bytes counts sampled once per second.

105 requests in 10.01s, 322 kB read
```